modify last_date calc logic

+++ download_stock_data.py
@@ -3,11 +3,12 @@ from datetime import datetime, date
 import os, sys
 
 def get_all_stock_data():
-    os.system('rm -rf marcap')
-    os.system('git clone "https://github.com/FinanceData/marcap.git" marcap')
+    if not os.path.isdir('marcap'):
+        os.system('rm -rf marcap')
+        os.system('git clone "https://github.com/FinanceData/marcap.git" marcap')
     from marcap import marcap_data
 
-    df = marcap_data('1995-05-02', datetime.today().strftime('%Y-%m-%d'))
+    df = marcap_data('2000-01-01', datetime.today().strftime('%Y-%m-%d'))
 
     return df
